+++ OrangeDesignSystem/Sources/OrangeDesignSystem/Modules/About/Configuration/Items/Recirculation/ODSRecirculationItem.swift
@@ -16,7 +16,7 @@ import SwiftUI
 /// Configuration to add in list the item that opens the apps recirculation view.
 ///
 /// By default, this item is placed between legal information item and rate the app item. That can be changed with new priority set in the configuration.
-public struct ODSMoreAppsItemConfig: ODSAboutListItemConfig {
+public struct ODSRecirculationItemConfig: ODSAboutListItemConfig {
 
     // =======================
     // MARK: Stored Properties
@@ -27,15 +27,6 @@ public struct ODSMoreAppsItemConfig: ODSAboutListItemConfig {
     public private(set) var target: ODSAboutListItemTarget
     public private(set) var priority: ODSAboutListItemPriority
 
-    // ==============
-    // MARK: - Source
-    // ==============
-
-    /// To define where the apps to display can be picked
-    public enum Source {
-        case remote(url: URL)
-        case local(path: URL)
-    }
 
     // =================
     // MARK: Initializer
@@ -44,29 +35,26 @@ public struct ODSMoreAppsItemConfig: ODSAboutListItemConfig {
     /// Initializes the configuration.
     ///
     /// - Parameters:
-    ///    - source: The source to use to get data
+    ///    - dataSource: The source to use to get data
     ///    - flattenApps: True if all apps must be palced in one list without categories, false (default) to keep categories
     ///    - cacheAppsIcons: True (default) to use app cache to save locally the apps stores icons, false otherwise
     ///    - enableHaptics: True (default) to enable vibrations with the module, false to disable
     ///    - priority: Priority to adjust the position of the item in the list.
-    public init(source: ODSMoreAppsItemConfig.Source,
+    public init(dataSource: ODSRecirculationDataSource,
                 flattenApps: Bool = false,
                 cacheAppsIcons: Bool = true,
                 enableHaptics: Bool = true,
                 priority: ODSAboutListItemPriority = .moreApps)
     {
-
-        var moreAppsView: MoreAppsView
-        switch source {
-        case let .remote(url):
-            moreAppsView = MoreAppsView(feedURL: url, flattenApps: flattenApps, cacheAppsIcons: cacheAppsIcons, enableHaptics: enableHaptics)
-        case let .local(path):
-            moreAppsView = MoreAppsView(localPath: path, flattenApps: flattenApps, cacheAppsIcons: cacheAppsIcons, enableHaptics: enableHaptics)
-        }
-
-        title = "modules.about.apps_recirculation.title".🌐
+        title = "modules.about.recirculation.title".🌐
         icon = Image("ic_mobile_apps", bundle: Bundle.ods)
-        target = .destination(AnyView(moreAppsView))
+        let destination = ODSRecirculation(
+            dataSource: dataSource,
+            flattenApps: flattenApps,
+            cacheAppsIcons: cacheAppsIcons,
+            enableHaptics: enableHaptics)
+
+        target = .destination(AnyView(destination))
         self.priority = priority
     }
 }